+++ src/main/java/io/github/dataforseo/client/model/SerpGoogleEventsLiveAdvancedResultInfo.java
@@ -34,7 +34,6 @@ import java.util.Set;
 import io.github.dataforseo.client.JSON;
 
 
-
 public class SerpGoogleEventsLiveAdvancedResultInfo  {
 
 
@@ -48,8 +47,7 @@ public class SerpGoogleEventsLiveAdvancedResultInfo  {
   }
 
   /**
-   * keyword received in a POST array
-* keyword is returned with decoded %## (plus character ‘+’ will be decoded to a space character)
+   * 
    * @return keyword
    */
   @javax.annotation.Nullable
@@ -95,7 +93,7 @@ public class SerpGoogleEventsLiveAdvancedResultInfo  {
   }
 
   /**
-   * search engine domain in a POST array
+   * 
    * @return seDomain
    */
   @javax.annotation.Nullable
@@ -118,7 +116,7 @@ public class SerpGoogleEventsLiveAdvancedResultInfo  {
   }
 
   /**
-   * location code in a POST array
+   * 
    * @return locationCode
    */
   @javax.annotation.Nullable
@@ -141,8 +139,7 @@ public class SerpGoogleEventsLiveAdvancedResultInfo  {
   }
 
   /**
-   * language code
-* in this case, the value will be null
+   * 
    * @return languageCode
    */
   @javax.annotation.Nullable
@@ -165,7 +162,7 @@ public class SerpGoogleEventsLiveAdvancedResultInfo  {
   }
 
   /**
-   * direct URL to search engine results
+   * 
    * @return checkUrl
    */
   @javax.annotation.Nullable
@@ -188,10 +185,7 @@ public class SerpGoogleEventsLiveAdvancedResultInfo  {
   }
 
   /**
-   * date and time when the result was received
-* in the UTC format: “yyyy-mm-dd hh-mm-ss +00:00”
-* example:
-* 2019-11-15 12:57:46 +00:00
+   * 
    * @return datetime
    */
   @javax.annotation.Nullable
@@ -214,8 +208,7 @@ public class SerpGoogleEventsLiveAdvancedResultInfo  {
   }
 
   /**
-   * autocorrection of the search engine
-* if the search engine provided results for a keyword that was corrected, we will specify the keyword corrected by the search engine and the type of autocorrection
+   * 
    * @return spell
    */
   @javax.annotation.Nullable
@@ -238,7 +231,7 @@ public class SerpGoogleEventsLiveAdvancedResultInfo  {
   }
 
   /**
-   * search refinement chips
+   * 
    * @return refinementChips
    */
   @javax.annotation.Nullable
@@ -261,9 +254,7 @@ public class SerpGoogleEventsLiveAdvancedResultInfo  {
   }
 
   /**
-   * types of search results found in SERP
-* possible item types:
-* event_item
+   * 
    * @return itemTypes
    */
   @javax.annotation.Nullable
@@ -278,48 +269,46 @@ public class SerpGoogleEventsLiveAdvancedResultInfo  {
 
   public static final String SERIALIZED_NAME_SE_RESULTS_COUNT = "se_results_count";
   @SerializedName(SERIALIZED_NAME_SE_RESULTS_COUNT)
-  private Long seResultsCount;
+  private Integer seResultsCount;
 
-  public SerpGoogleEventsLiveAdvancedResultInfo seResultsCount(Long seResultsCount) {
+  public SerpGoogleEventsLiveAdvancedResultInfo seResultsCount(Integer seResultsCount) {
     this.seResultsCount = seResultsCount;
     return this;
   }
 
   /**
-   * total number of results in SERP
-* in this case, the value will be 0
-* this search engine does not indicate the total number of results
+   * 
    * @return seResultsCount
    */
   @javax.annotation.Nullable
-  public Long getSeResultsCount() {
+  public Integer getSeResultsCount() {
     return seResultsCount;
   }
 
-  public void setSeResultsCount(Long seResultsCount) {
+  public void setSeResultsCount(Integer seResultsCount) {
     this.seResultsCount = seResultsCount;
   }
 
 
   public static final String SERIALIZED_NAME_ITEMS_COUNT = "items_count";
   @SerializedName(SERIALIZED_NAME_ITEMS_COUNT)
-  private Long itemsCount;
+  private Integer itemsCount;
 
-  public SerpGoogleEventsLiveAdvancedResultInfo itemsCount(Long itemsCount) {
+  public SerpGoogleEventsLiveAdvancedResultInfo itemsCount(Integer itemsCount) {
     this.itemsCount = itemsCount;
     return this;
   }
 
   /**
-   * the number of results returned in the items array
+   * 
    * @return itemsCount
    */
   @javax.annotation.Nullable
-  public Long getItemsCount() {
+  public Integer getItemsCount() {
     return itemsCount;
   }
 
-  public void setItemsCount(Long itemsCount) {
+  public void setItemsCount(Integer itemsCount) {
     this.itemsCount = itemsCount;
   }
 
@@ -334,7 +323,7 @@ public class SerpGoogleEventsLiveAdvancedResultInfo  {
   }
 
   /**
-   * items of the element
+   * 
    * @return items
    */
   @javax.annotation.Nullable
