+++ src/main/java/io/github/dataforseo/client/model/SerpGoogleFinanceExploreLiveAdvancedResultInfo.java
@@ -34,7 +34,6 @@ import java.util.Set;
 import io.github.dataforseo.client.JSON;
 
 
-
 public class SerpGoogleFinanceExploreLiveAdvancedResultInfo  {
 
 
@@ -48,8 +47,7 @@ public class SerpGoogleFinanceExploreLiveAdvancedResultInfo  {
   }
 
   /**
-   * keyword received in a POST array
-* the keyword is returned with decoded %## (plus character ‘+’ will be decoded to a space character)
+   * 
    * @return keyword
    */
   @javax.annotation.Nullable
@@ -95,7 +93,7 @@ public class SerpGoogleFinanceExploreLiveAdvancedResultInfo  {
   }
 
   /**
-   * search engine domain in a POST array
+   * 
    * @return seDomain
    */
   @javax.annotation.Nullable
@@ -118,7 +116,7 @@ public class SerpGoogleFinanceExploreLiveAdvancedResultInfo  {
   }
 
   /**
-   * location code in a POST array
+   * 
    * @return locationCode
    */
   @javax.annotation.Nullable
@@ -141,7 +139,7 @@ public class SerpGoogleFinanceExploreLiveAdvancedResultInfo  {
   }
 
   /**
-   * language code in a POST array
+   * 
    * @return languageCode
    */
   @javax.annotation.Nullable
@@ -164,8 +162,7 @@ public class SerpGoogleFinanceExploreLiveAdvancedResultInfo  {
   }
 
   /**
-   * direct URL to search engine results
-* you can use it to make sure that we provided accurate results
+   * 
    * @return checkUrl
    */
   @javax.annotation.Nullable
@@ -188,10 +185,7 @@ public class SerpGoogleFinanceExploreLiveAdvancedResultInfo  {
   }
 
   /**
-   * date and time when the result was received
-* in the UTC format: “yyyy-mm-dd hh-mm-ss +00:00”
-* example:
-* 2019-11-15 12:57:46 +00:00
+   * 
    * @return datetime
    */
   @javax.annotation.Nullable
@@ -214,8 +208,7 @@ public class SerpGoogleFinanceExploreLiveAdvancedResultInfo  {
   }
 
   /**
-   * autocorrection of the search engine
-* if the search engine provided results for a keyword that was corrected, we will specify the keyword corrected by the search engine and the type of autocorrection
+   * 
    * @return spell
    */
   @javax.annotation.Nullable
@@ -238,8 +231,7 @@ public class SerpGoogleFinanceExploreLiveAdvancedResultInfo  {
   }
 
   /**
-   * search refinement chips
-* in this case, the value will be null
+   * 
    * @return refinementChips
    */
   @javax.annotation.Nullable
@@ -262,9 +254,7 @@ public class SerpGoogleFinanceExploreLiveAdvancedResultInfo  {
   }
 
   /**
-   * types of search results in SERP
-* contains types of search results (items) found in SERP;
-* possible item types: google_finance_hero_groups, google_finance_interested, google_finance_news, google_finance_earnings_calendar, google_finance_most_followed, google_finance_market_trends, google_finance_people_also_search
+   * 
    * @return itemTypes
    */
   @javax.annotation.Nullable
@@ -279,46 +269,46 @@ public class SerpGoogleFinanceExploreLiveAdvancedResultInfo  {
 
   public static final String SERIALIZED_NAME_SE_RESULTS_COUNT = "se_results_count";
   @SerializedName(SERIALIZED_NAME_SE_RESULTS_COUNT)
-  private Long seResultsCount;
+  private Integer seResultsCount;
 
-  public SerpGoogleFinanceExploreLiveAdvancedResultInfo seResultsCount(Long seResultsCount) {
+  public SerpGoogleFinanceExploreLiveAdvancedResultInfo seResultsCount(Integer seResultsCount) {
     this.seResultsCount = seResultsCount;
     return this;
   }
 
   /**
-   * total number of results in SERP
+   * 
    * @return seResultsCount
    */
   @javax.annotation.Nullable
-  public Long getSeResultsCount() {
+  public Integer getSeResultsCount() {
     return seResultsCount;
   }
 
-  public void setSeResultsCount(Long seResultsCount) {
+  public void setSeResultsCount(Integer seResultsCount) {
     this.seResultsCount = seResultsCount;
   }
 
 
   public static final String SERIALIZED_NAME_ITEMS_COUNT = "items_count";
   @SerializedName(SERIALIZED_NAME_ITEMS_COUNT)
-  private Long itemsCount;
+  private Integer itemsCount;
 
-  public SerpGoogleFinanceExploreLiveAdvancedResultInfo itemsCount(Long itemsCount) {
+  public SerpGoogleFinanceExploreLiveAdvancedResultInfo itemsCount(Integer itemsCount) {
     this.itemsCount = itemsCount;
     return this;
   }
 
   /**
-   * the number of results returned in the items array
+   * 
    * @return itemsCount
    */
   @javax.annotation.Nullable
-  public Long getItemsCount() {
+  public Integer getItemsCount() {
     return itemsCount;
   }
 
-  public void setItemsCount(Long itemsCount) {
+  public void setItemsCount(Integer itemsCount) {
     this.itemsCount = itemsCount;
   }
 
@@ -333,8 +323,7 @@ public class SerpGoogleFinanceExploreLiveAdvancedResultInfo  {
   }
 
   /**
-   * market indexes related to the market trends element
-* possible type of items: google_finance_asset_pair_element, google_finance_market_instrument_element, google_finance_market_index_element
+   * 
    * @return items
    */
   @javax.annotation.Nullable
